requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
odmantic>=1.0.0
python-dotenv>=1.0.0
PyJWT>=2.8.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4